requirements.txt
IMDbPY==2022.7.9
openai==0.19.0